+++ src/renderer/src/components/ModelLibrary/providers/index.ts
@@ -8,6 +8,7 @@ export * from './DeepseekParameters';
 export * from './GeminiParameters';
 export * from './GpustackParameters';
 export * from './GroqParameters';
+export * from './KimiPlanParameters';
 export * from './LitellmParameters';
 export * from './MinimaxParameters';
 export * from './LmStudioParameters';